Update stories resolution
