+++ tools/storybook-utils/tsconfig.json
@@ -5,10 +5,7 @@
     "outDir": "dist",
     "rootDir": "src",
     "baseUrl": ".",
-    "paths": {
-      "@leafygreen-ui/icon/dist/*": ["../icon/src/generated/*"],
-      "@leafygreen-ui/*": ["../*/src"]
-    }
+    "paths": {}
   },
   "include": [
     "src/**/*"
